requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.3
google-api-python-client==2.104.0
beautifulsoup4==4.12.2
yt-dlp==2023.11.16